--- code/algos.py
@@ -1,4 +1,0 @@
-## algos.py
-from SVR_permutation import SVR_permutation
-from MARS_permutation import MARS_permutation
-from RandomForest_permutation import RF_permutation
