+++ code/run_experiments.py
@@ -6,7 +6,9 @@ from sklearn.model_selection import GridSearchCV
 from sklearn.model_selection import KFold
 import colorama
 from termcolor import colored
-import algos
+from algorithms.SVR_permutation import SVR_permutation
+from algorithms.MARS_permutation import MARS_permutation
+from algorithms.RandomForest_permutation import RF_permutation
 
 colorama.init()
 
@@ -59,15 +61,15 @@ algo_args = (args.traitData,
 
 print("Beginning permutation testing...")
 if "SVR" in args.algo:
-    algos.SVR_permutation(*algo_args)
+    SVR_permutation(*algo_args)
     print(colored("SVR complete.", 'green'))
 
 if "MARS" in args.algo:
-    algos.MARS_permutation(*algo_args)
+    MARS_permutation(*algo_args)
     print(colored('MARS complete.', 'green'))
 
 if "RF" in args.algo:
-    algos.RF_permutation(*algo_args)
+    RF_permutation(*algo_args)
     print(colored("RF complete.", 'green'))
 
 if "Linear" in args.algo:
